backend/requirements.txt
fastapi==0.115.4
PyJWT==2.9.0
PyJWT==2.9.0
python-dotenv==1.0.1
sqlmodel==0.0.22
psycopg2-binary==2.9.10